# Specialized dot product