--- a/src/makeBBBB.jl
+++ b/src/makeBBBB.jl
@@ -21,22 +21,48 @@ end
 
 #########################
 
-function makeBBBBs(n,nlmi,A,AA,myA,W,to,qA,sigmaA)
+function makeBBBBs(n,nlmi,A,AA,W,to,qA,sigmaA)
     BBBB = zeros(Float64, n, n)
     @inbounds for ilmi = 1:nlmi
         Wilmi = W[ilmi]
         AAilmi = AA[ilmi]
         Ailmi = A[ilmi,:]
         @timeit to "BBBBs" begin
-        BBBB += makeBBBBsi(ilmi,Ailmi,AAilmi,myA,Wilmi,n,to,qA,sigmaA)
+        BBBB += makeBBBBsi(ilmi,Ailmi,AAilmi,Wilmi,n,to,qA,sigmaA)
         end
     end
 
     return BBBB
 end
 
+# Computes `⟨A * W, W * B⟩` for symmetric sparse matrices `A` and `B`
+function _dot(A::SparseMatrixCSC, B::SparseMatrixCSC, W::Matrix)
+    @assert LinearAlgebra.checksquare(W) == LinearAlgebra.checksquare(A) == LinearAlgebra.checksquare(B)
+    result = zero(eltype(A))
+    for i in axes(A, 2)
+        nzA = nzrange(A, i)
+        if !isempty(nzA)
+            for j in axes(B, 2)
+                nzB = nzrange(B, j)
+                if !isempty(nzB)
+                    AW = zero(result)
+                    for k in nzA
+                        AW += nonzeros(A)[k] * W[rowvals(A)[k], j]
+                    end
+                    WB = zero(result)
+                    for k in nzB
+                        WB += W[i, rowvals(B)[k]] * nonzeros(B)[k]
+                    end
+                    result += AW * WB
+                end
+            end
+        end
+    end
+    return result
+end
+
 #####
-function makeBBBBsi(ilmi,Ailmi,AAilmi,myA,Wilmi::Matrix{T},n,to,qA,sigmaA) where {T}
+function makeBBBBsi(ilmi,Ailmi,AAilmi,Wilmi::Matrix{T},n,to,qA,sigmaA) where {T}
     BBBB = zeros(T, n, n)
     tmp1 = Matrix{T}(undef,size(Wilmi, 2), size(Ailmi[1], 1))
     # tmp = Matrix{Float64}(undef,size(Wilmi, 1), size(Wilmi, 1))
@@ -81,24 +107,22 @@ function makeBBBBsi(ilmi,Ailmi,AAilmi,myA,Wilmi::Matrix{T},n,to,qA,sigmaA) where
                 mul!(tmp1,Ailmi[i+1],Wilmi)
                 @inbounds for jj = ii:n
                     j = sigmaA[jj,ilmi]
-                    if ~isempty(myA[ilmi1+j].iind)
-                        myAjjj = myA[ilmi1+j]
-                        iii_j = myAjjj.iind
-                        jjj_j = myAjjj.jind
-                        vvv_j = myAjjj.nzval
+                    Ajjj = Ailmi[j+1]
+                    if !iszero(nnz(Ajjj))
                         ttt = 0.0
                         # @timeit to "BBBBtwo_i" begin
-                        @inbounds for iAj in eachindex(iii_j)
+                        @inbounds for jjjjAj in axes(Ajjj, 2)
+                            for k in nzrange(Ajjj, jjjjAj)
                             # @timeit to "BBBBtwo_ii_A" begin
-                            iiijAj = iii_j[iAj]
-                            jjjjAj = jjj_j[iAj]
+                                iiijAj = rowvals(Ajjj)[k]
                             # end
                             # vvvj = -vvv_j[iAj]    
                             # @timeit to "BBBBtwo_i_B" begin
-                            ttt1 = dot(tmp1[:,iiijAj],Wilmi[:,jjjjAj])
+                                ttt1 = dot(tmp1[:,iiijAj],Wilmi[:,jjjjAj])
                             # end
                             # @timeit to "BBBBtwo_i_C" begin
-                            ttt -= ttt1 * vvv_j[iAj]
+                                ttt += ttt1 * nonzeros(Ajjj)[k]
+                            end
                             # end
                         end
                         # end
@@ -115,31 +139,24 @@ function makeBBBBsi(ilmi,Ailmi,AAilmi,myA,Wilmi::Matrix{T},n,to,qA,sigmaA) where
                 # @show "three"
                 # @show ilmi
                 # @show ii
-                if ~isempty(myA[ilmi1+i].iind)
-                    myAiii = myA[ilmi1+i]
-                    if size(myAiii.iind,1) > 1
+                if !iszero(nnz(Ailmi[i+1]))
+                    if nnz(Ailmi[i+1]) > 1
                         # @timeit to "BBBBthree>1" begin
-                        iii_is = union(myAiii.iind)
                         # @show iii_i,myAiii.jind
                         # iii_is = iii_i[1:Int64(sqrt(length(iii_i)))]
                         # jjj_i = myAiii.jind
                         # vvv_i = myAiii.nzval
-                        mya1 = Ailmi[i+1][iii_is,iii_is]
                         @inbounds for jj = ii:n
                             j = sigmaA[jj,ilmi]
-                            if ~isempty(myA[ilmi1+j].iind)
+                            if !iszero(nnz(Ailmi[j+1]))
                                 # @timeit to "BBBBthree_1>1" begin
-                                myAjjj = myA[ilmi1+j]
-                                iii_js = union(myAjjj.iind)
                                 # iii_js = iii_j[1:Int64(sqrt(length(iii_j)))]
                                 # jjj_j = myAjjj.jind
                                 # vvv_j = myAjjj.nzval
                                 # ttt = 0.0
-                                myw = Wilmi[iii_is,iii_js]                            
-                                mya2 = Ailmi[j+1][iii_js,iii_js]
                                 # end
                                 # @timeit to "BBBBthree_2>1" begin
-                                ttt = transpose(vec(transpose(mya1 * myw))) * vec(mya2 * transpose(myw))
+                                ttt = _dot(Ailmi[i+1], Ailmi[j+1], Wilmi)
                                 # end
                                 # @inbounds for iAj in eachindex(iii_j)
                                 #     ttt1 = 0.0
@@ -167,16 +184,18 @@ function makeBBBBsi(ilmi,Ailmi,AAilmi,myA,Wilmi::Matrix{T},n,to,qA,sigmaA) where
                         end   
                     else
                         @timeit to "BBBBthree=1" begin
-                        iiiiAi = myAiii.iind[1]
-                        jjjiAi = myAiii.jind[1]
-                        vvvi = myAiii.nzval[1]
+                        # A is symmetric
+                        iiiiAi = jjjiAi = only(rowvals(Ailmi[i+1]))
+                        vvvi = only(nonzeros(Ailmi[i+1]))
                         @inbounds for jj = ii:n
                             j = sigmaA[jj,ilmi]
-                            if ~isempty(myA[ilmi1+j].iind)
-                                myAjjj = myA[ilmi1+j]
-                                iiijAj = myAjjj.iind[1]
-                                jjjjAj = myAjjj.jind[1]
-                                vvvj = myAjjj.nzval[1]
+                            Ajjj = Ailmi[j+1]
+                            # As we sort the matrices in decreasing `nnz` order,
+                            # the rest of matrices is either zero or have only
+                            # one entry
+                            if !iszero(nnz(Ajjj))
+                                iiijAj = jjjjAj = only(rowvals(Ajjj))
+                                vvvj = only(nonzeros(Ajjj))
                                 ttt = vvvi * Wilmi[iiiiAi,iiijAj] * Wilmi[jjjiAi,jjjjAj] * vvvj
                                 if i >= j
                                     BBBB[i,j] = ttt