Allocating the submatrices and the result of their product was taking a lot of time.
By writing a custom `_dot` function that is allocation-free, we gain a lot in term of perf and allocations and we don't need `myA` anymore.
The result of `BenchmarkTools.@benchmark optimize!(model)` (using `0` for `verb`) on `examples/solve_sdpa.jl` gives:
Before the PR:
```
BenchmarkTools.Trial: 16 samples with 1 evaluation per sample.
 Range (min … max):  285.593 ms … 365.039 ms  ┊ GC (min … max): 0.00% … 14.98%
 Time  (median):     326.396 ms               ┊ GC (median):    8.39%
 Time  (mean ± σ):   326.294 ms ±  19.043 ms  ┊ GC (mean ± σ):  8.00% ±  4.54%

  ▁     ▁               ▁    ██  █ █ ▁   ▁    ▁    ▁          ▁  
  █▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁██▁▁█▁█▁█▁▁▁█▁▁▁▁█▁▁▁▁█▁▁▁▁▁▁▁▁▁▁█ ▁
  286 ms           Histogram: frequency by time          365 ms <

 Memory estimate: 131.18 MiB, allocs estimate: 2156231.
```
After the PR:
```
BenchmarkTools.Trial: 49 samples with 1 evaluation per sample.
 Range (min … max):   87.494 ms … 146.038 ms  ┊ GC (min … max): 0.00% … 3.78%
 Time  (median):      99.270 ms               ┊ GC (median):    4.47%
 Time  (mean ± σ):   103.155 ms ±  13.001 ms  ┊ GC (mean ± σ):  3.84% ± 3.64%

  ▅    ▂ ▅   ▅█                                                  
  █▁▅▅██▁██▅███▅▅▁█▅▁▁▁▅▁▅█▁▅▅▁▁▅█▁▅▁▅▁▅▁▁▁▁▁▁▁▅▁▁▅▁▁▁▁▁▁▁▁▁▁▁▅ ▁
  87.5 ms          Histogram: frequency by time          146 ms <

 Memory estimate: 37.39 MiB, allocs estimate: 69245.
```
So 96% of reduction in number of allocations, 70% of reduction both in memory allocated and time