@kocvara 2025-05-23
Great, thanks! Yes, myA was a Matlab-style relict which I didn’t know how to write better… M.

> On 23 May 2025, at 07:07, Benoît Legat ***@***.***> wrote:
> 
> Allocating the submatrices and the result of their product was taking a lot of time.
> By writing a custom _dot function that is allocation-free, we gain a lot in term of perf and allocations and we don't need myA anymore.
> The result of ***@***.*** optimize!(model) on examples/solve_sdpa.jl gives:
> Before the PR:
> 
> BenchmarkTools.Trial: 16 samples with 1 evaluation per sample.
>  Range (min … max):  285.593 ms … 365.039 ms  ┊ GC (min … max): 0.00% … 14.98%
>  Time  (median):     326.396 ms               ┊ GC (median):    8.39%
>  Time  (mean ± σ):   326.294 ms ±  19.043 ms  ┊ GC (mean ± σ):  8.00% ±  4.54%
> 
>   ▁     ▁               ▁    ██  █ █ ▁   ▁    ▁    ▁          ▁  
>   █▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁██▁▁█▁█▁█▁▁▁█▁▁▁▁█▁▁▁▁█▁▁▁▁▁▁▁▁▁▁█ ▁
>   286 ms           Histogram: frequency by time          365 ms <
> 
>  Memory estimate: 131.18 MiB, allocs estimate: 2156231.
> After the PR:
> 
> BenchmarkTools.Trial: 49 samples with 1 evaluation per sample.
>  Range (min … max):   87.494 ms … 146.038 ms  ┊ GC (min … max): 0.00% … 3.78%
>  Time  (median):      99.270 ms               ┊ GC (median):    4.47%
>  Time  (mean ± σ):   103.155 ms ±  13.001 ms  ┊ GC (mean ± σ):  3.84% ± 3.64%
> 
>   ▅    ▂ ▅   ▅█                                                  
>   █▁▅▅██▁██▅███▅▅▁█▅▁▁▁▅▁▅█▁▅▅▁▁▅█▁▅▁▅▁▅▁▁▁▁▁▁▁▅▁▁▅▁▁▁▁▁▁▁▁▁▁▁▅ ▁
>   87.5 ms          Histogram: frequency by time          146 ms <
> 
>  Memory estimate: 37.39 MiB, allocs estimate: 69245.
> So 96% of reduction in number of allocations, 70% of reduction both in memory allocated and time
> 
> You can view, comment on, or merge this pull request online at:
> 
>   https://github.com/kocvara/Loraine.jl/pull/25
> 
> Commit Summary
> 
> a953ea3 <https://github.com/kocvara/Loraine.jl/pull/25/commits/a953ea392f87d40f5d9a17a964635abfa5817660> Specialized dot product
> File Changes (4 files <https://github.com/kocvara/Loraine.jl/pull/25/files>)
> M src/Solvers.jl <https://github.com/kocvara/Loraine.jl/pull/25/files#diff-407709cbf10fa8ba497bb90a66f1eaa3021559ef448f32694315a47032e5b62b> (1)
> M src/makeBBBB.jl <https://github.com/kocvara/Loraine.jl/pull/25/files#diff-89aeffd260544e25ec5244062af35089d43b2efb1e26a3798ddeefa0f40474dc> (83)
> M src/model.jl <https://github.com/kocvara/Loraine.jl/pull/25/files#diff-4439b24aa0d88d7b848e34d6318dcba50ea599b7ef9ab37f879d68cbd6c7023b> (14)
> M src/predictor_corrector.jl <https://github.com/kocvara/Loraine.jl/pull/25/files#diff-74e1286281191acc0f86aa4505b6d46f51a3e5c04e027ab979778cd01c99a8b2> (2)
> Patch Links:
> 
> https://github.com/kocvara/Loraine.jl/pull/25.patch
> https://github.com/kocvara/Loraine.jl/pull/25.diff
> —
> Reply to this email directly, view it on GitHub <https://github.com/kocvara/Loraine.jl/pull/25>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/AFIXXOV3K46TNYBADKFEODT2723LXAVCNFSM6AAAAAB5X234MOVHI2DSMVQWIX3LMV43ASLTON2WKOZTGA4DKMRZGM3DEMY>.
> You are receiving this because you are subscribed to this thread.
>